Remove type module to try and fix vitest error
